Add diff-pairs task

+++ diff-pairs/diff_pairs.h
@@ -8,10 +8,10 @@
 #include <set>
 
 int64_t CountPairs(const std::vector<int>& data, int x) {
-    std::map<int, int> freq;
-    int ans = 0;
+    std::map<int64_t, int64_t> freq;
+    int64_t ans = 0;
     for (size_t i = 0; i < data.size(); ++i) {
-        int complement = x - data[i];
+        int64_t complement = x - data[i];
         if (freq.contains(complement)) {
             ans += freq[complement];
         }
